Rename AuctionHouseRoutes to AuctionHouseService
Move class to package service.
