+++ src/test/scala/com/ion/trials/akka/util/AuctionTestData.scala
@@ -9,7 +9,7 @@ import com.ion.trials.akka.auction.AuctionTypes.{
   AuctionRule
 }
 import com.ion.trials.akka.auction.Planned.plannedStr
-import com.ion.trials.akka.routes.{AuctionRuleParams, AuctionRuleParamsUpdate}
+import com.ion.trials.akka.service.{AuctionRuleParams, AuctionRuleParamsUpdate}
 
 import scala.concurrent.duration._
 import scala.language.postfixOps
